requirements.txt
av
basicsr @ git+https://github.com/XPixelGroup/BasicSR.git@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
facexlib>=0.2.5
ffmpeg-python
gfpgan>=1.3.5
ipykernel
loguru
numpy==1.24.3
opencv-python
opencv-python-headless
Pillow
pims
pims
pyqtgraph
PySide6
pyzmq
torch==1.13.1
# torchvision==0.14
tqdm
zmq